+++ src/bl64-ui-env.bash
@@ -4,9 +4,9 @@
 
 # shellcheck disable=SC2034
 {
-  declare BL64_UI_VERSION='2.0.1'
+  declare BL64_UI_VERSION='2.0.2'
 
   declare BL64_UI_MODULE='0'
 
-  declare BL64_UI_READ_TIMEOUT='60'
+  declare BL64_UI_CONFIRMATION_TIMEOUT='60'
 }