+++ src/bl64-ui.bash
@@ -19,12 +19,12 @@
 #######################################
 function bl64_ui_ask_confirmation() {
   bl64_dbg_lib_show_function "$@"
-  local question="${1:-Please confirm the operation by writting the message}"
+  local question="${1:-Please type in the confirmation message to proceed}"
   local confirmation="${2:-confirm-operation}"
   local input=''
 
   bl64_msg_show_input "${question} [${confirmation}]: "
-  read -r -t "$BL64_UI_READ_TIMEOUT" input
+  read -r -t "$BL64_UI_CONFIRMATION_TIMEOUT" input
 
   if [[ "$input" != "$confirmation" ]]; then
     bl64_msg_show_error 'confirmation verification failed'