+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [20.11.0]
+
+### Added
+
+- fmt
+  - bl64_fmt_version_is_major_minor
+  - bl64_fmt_version_is_major
+  - bl64_fmt_version_is_semver
+
+### Changed
+
+- msg
+  - bl64_msg_show_batch_start: default msg is now BL64_SCRIPT_ID
+  - bl64_msg_show_batch_stop: default msg is now BL64_SCRIPT_ID
+
+### Fixed
+
+- core
+  - Missing OS in compatibility list
+
 ## [20.10.0]
 
 ### Added